Fix rebase issue

+++ src/engine/sim_engine/keys.rs
@@ -78,7 +78,7 @@ impl KeyActions for SimKeyActions {
                 ErrorEnum::Invalid,
                 format!(
                     "Provided key exceeded maximum allow length of {}",
-                    Bytes(MAX_STRATIS_PASS_SIZE as u64)
+                    Bytes(MAX_STRATIS_PASS_SIZE as u128)
                 ),
             ));
         }
