+++ src/engine/strat_engine/keys.rs
@@ -397,7 +397,7 @@ impl KeyActions for StratKeyActions {
                 ErrorEnum::Invalid,
                 format!(
                     "Provided key exceeded maximum allow length of {}",
-                    Bytes(MAX_STRATIS_PASS_SIZE as u64)
+                    Bytes(MAX_STRATIS_PASS_SIZE as u128)
                 ),
             ));
         }
